Restructure model modules
